+++ examples/limit.rb
@@ -8,7 +8,7 @@
 require "rubygems"
 require "highline/import"
 
-puts "Using: #{$terminal.terminal.class}"
+puts "Using: #{HighLine.default_instance.terminal.class}"
 puts
 
 text = ask("Enter text (max 10 chars): ") { |q| q.limit = 10 }