+++ examples/overwrite.rb
@@ -8,7 +8,7 @@
 require 'rubygems'
 require 'highline/import'
 
-puts "Using: #{$terminal.terminal.class}"
+puts "Using: #{HighLine.default_instance.terminal.class}"
 puts
 
 prompt = "here is your password:"