+++ lib/highline/import.rb
@@ -10,22 +10,19 @@
 require "highline"
 require "forwardable"
 
-$terminal = HighLine.new
-
 #
 # <tt>require "highline/import"</tt> adds shortcut methods to Kernel, making
 # {HighLine#agree}, {HighLine#ask}, {HighLine#choose} and {HighLine#say}
 # globally available.  This is handy for
-# quick and dirty input and output.  These methods use the HighLine object in
-# the global variable <tt>$terminal</tt>, which is initialized to use
-# <tt>$stdin</tt> and <tt>$stdout</tt> (you are free to change this).
-# Otherwise, these methods are identical to their {HighLine} counterparts, see that
-# class for detailed explanations.
+# quick and dirty input and output.  These methods use HighLine.default_instance
+# which is initialized to use <tt>$stdin</tt> and <tt>$stdout</tt> (you are free
+# to change this).
+# Otherwise, these methods are identical to their {HighLine} counterparts,
+# see that class for detailed explanations.
 #
 module Kernel
   extend Forwardable
-  def_delegators :$terminal, :agree, :ask, :choose, :say,
-                 :use_color=, :use_color?, :reset_use_color
+  def_instance_delegators :HighLine, :agree, :ask, :choose, :say
 end
 
 # When requiring 'highline/import' HighLine adds {#or_ask} to Object so