+++ test/test_import.rb
@@ -20,13 +20,21 @@ class TestImport < Minitest::Test
     assert_respond_to(self, :choose)
     assert_respond_to(self, :say)
   end
+
+  def test_healthy_default_instance_after_import
+    refute_nil HighLine.default_instance
+    assert_instance_of HighLine, HighLine.default_instance
+
+    # If correctly initialized, it will contain several ins vars.
+    refute_empty HighLine.default_instance.instance_variables
+  end
   
   def test_or_ask
-    old_terminal = $terminal
+    old_instance = HighLine.default_instance
     
     input     = StringIO.new
     output    = StringIO.new
-    $terminal = HighLine.new(input, output)  
+    HighLine.default_instance = HighLine.new(input, output)  
     
     input << "10\n"
     input.rewind
@@ -40,16 +48,16 @@ class TestImport < Minitest::Test
     
     assert_equal(10, 20.or_ask("How much?  ", Integer) { |q| q.in = 1..10 })
   ensure
-    $terminal = old_terminal
+    HighLine.default_instance = old_instance
   end
   
   def test_redirection
-    old_terminal = $terminal
+    old_instance = HighLine.default_instance
     
-    $terminal = HighLine.new(nil, (output = StringIO.new))
+    HighLine.default_instance = HighLine.new(nil, (output = StringIO.new))
     say("Testing...")
     assert_equal("Testing...\n", output.string)
   ensure
-    $terminal = old_terminal
+    HighLine.default_instance = old_instance
   end
 end